requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Phase 2 — OPC-UA simulator + collector
asyncua==1.1.5
structlog==24.1.0

# Phase 3 — JWT auth + alerting
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
